Double instead of int on every stats input
* Double instead of int on every stats input
* Additional implementation of input parameters serialization

+++ DamageCalculator/src/damage_calculator/InputParameters.java
@@ -11,20 +11,20 @@ public class InputParameters {
 
     public int maxTier = 10;
 
-    public int troopAttack = 100;
-    public int infantryAttack = 100;
-    public int cavalryAttack = 100;
-    public int distanceAttack = 100;
-    public int artilleryAttack = 100;
+    public double troopAttack = 100;
+    public double infantryAttack = 100;
+    public double cavalryAttack = 100;
+    public double distanceAttack = 100;
+    public double artilleryAttack = 100;
 
-    public int troopDamage = 100;
-    public int infantryDamage = 100;
-    public int cavalryDamage = 100;
-    public int distanceDamage = 100;
-    public int artilleryDamage = 100;
+    public double troopDamage = 100;
+    public double infantryDamage = 100;
+    public double cavalryDamage = 100;
+    public double distanceDamage = 100;
+    public double artilleryDamage = 100;
 
-    public int distanceVsInfantryDamage = 100;
-    public int cavalryVsInfantryDamage = 100;
+    public double distanceVsInfantryDamage = 100;
+    public double cavalryVsInfantryDamage = 100;
 
     @Override
     public String toString() {
